// apps/front-end/src/shared/api/generated/model/clientMetricDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Capital CRM API
 * Documentação da API do Capital CRM Platform
 * OpenAPI spec version: 1.0
 */

export interface ClientMetricDto {
  /** ID do cliente associado à métrica */
  clientId: string;
  /** Número de visualizações do cliente */
  views: number;
  /** Data da última visualização do cliente */
  lastViewedAt: string;
}