// apps/front-end/src/shared/api/generated/model/clientResponseDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Capital CRM API
 * Documentação da API do Capital CRM Platform
 * OpenAPI spec version: 1.0
 */

export interface ClientResponseDto {
  /** ID do cliente */
  id: string;
  /** Nome do cliente */
  name: string;
  /** Salário do cliente */
  salary: number;
  /** Valor da empresa do cliente */
  companyValue: number;
  /** Define se o cliente está selecionado */
  isSelected?: boolean;
  /** Data de criação do cliente */
  createdAt: string;
  /** Data de atualização do cliente */
  updatedAt: string;
}