// apps/front-end/src/shared/api/generated/model/createClientDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Capital CRM API
 * Documentação da API do Capital CRM Platform
 * OpenAPI spec version: 1.0
 */

export interface CreateClientDto {
  /** Nome do cliente */
  name: string;
  /** Salário do cliente */
  salary: number;
  /** Valor da empresa do cliente */
  companyValue: number;
  /** Define se o cliente está selecionado */
  isSelected?: boolean;
}