// apps/front-end/src/shared/api/generated/model/dashboardResponseDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Capital CRM API
 * Documentação da API do Capital CRM Platform
 * OpenAPI spec version: 1.0
 */
import type { TopClientDto } from './topClientDto';

export interface DashboardResponseDto {
  /** Total de clientes ativos na base */
  totalClients: number;
  /** Soma do valor de mercado das empresas */
  totalCompanyValue: number;
  /** Média salarial dos clientes */
  averageSalary: number;
  /** Top 5 clientes mais visualizados */
  topViewedClients: TopClientDto[];
}