// apps/front-end/src/shared/api/generated/model/loginDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Capital CRM API
 * Documentação da API do Capital CRM Platform
 * OpenAPI spec version: 1.0
 */

export interface LoginDto {
  /** E-mail do usuário */
  email: string;
  /**
   * Senha do usuário
   * @minLength 6
   */
  password: string;
}